#50: Rename rd_class.py to rd_object.py.
