+++ tests/unit/test_raindrop_processor.py
